Rewriting to use taglib rather than id3lib.
The use of id3lib to handle parsing and modifying the ID3 headers
of the MP3 files was causing album artwork to be lost as well
as updating the version of the ID3 specification being used.

Switching to taglib seems to eliminate these problems and, most
importantly, retains the album artwork.